\name{ff-MAGMA testing scripts}
\alias{ff-MAGMA.test}

\title{
ff-MAGMA testing scipts
}

\description{
Scripts used for testing various functions in ff-MAGMA.
}

\references{
http://cran.r-project.org/web/packages/fields/fields.pdf
http://www.image.ucar.edu/~nychka/Fields/
http://icl.cs.utk.edu/magma/
}

\author{
John Paige
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{fields}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ spatial statistics }
